+++ docs/source/changelog.rst
@@ -11,6 +11,7 @@ Change Log
   They're still available on their old location for backwards compatibility for now.
   Ofcourse, they're also still on their old "shortcut" location in ``Pyro4`` directly.
 - Removed the publicly visible serializer id numbers from the message module. They're internal protocol details, user code should always refer to serializers by their name.
+- When a connection cannot be made, the address Pyro tries to connect to is now also included in the error message.
 - Fixed various examples.